Safely encode temporary branch name
The temporary branch name created for the source of the pull request
uses the submitter name, which may contain invalid characters. Use
the "slugify" (adapted from Django) method to get a valid encoding
that preserves as much as possible of the original name, and fall
back to base64-encoding if that fails.